pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "googletrans==4.0.0rc1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pdf2image>=1.17.0",
    "pillow>=11.2.1",
    "pytesseract>=0.3.13",
    "streamlit>=1.45.1",
]
